Refactored KeyValueStore to be itself IKeyValueStore
