+++ Nitrox.Test/Model/Helper/KeyValueStoreTest.cs
@@ -1,28 +1,23 @@
-using System;
-using System.Threading.Tasks;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using NitroxModel.Platforms.OS.Windows.Internal;
 
-namespace NitroxModel.Helper
+namespace NitroxModel.Helper;
+
+[TestClass]
+public class KeyValueStoreTest
 {
-    [TestClass]
-    public class KeyValueStoreTest
+    [TestMethod]
+    public void SetAndReadValue()
     {
-        [TestMethod]
-        public void SetAndReadValue()
-        {
-            const string TEST_KEY = "test";
+        const string TEST_KEY = "test";
 
-            KeyValueStore.SetValue<int>(TEST_KEY, -50);
-            Assert.AreEqual(-50, KeyValueStore.GetValue<int>(TEST_KEY));
+        KeyValueStore.Instance.SetValue(TEST_KEY, -50);
+        Assert.AreEqual(-50, KeyValueStore.Instance.GetValue<int>(TEST_KEY));
 
-            KeyValueStore.SetValue<int>(TEST_KEY, 1337);
-            Assert.AreEqual(1337, KeyValueStore.GetValue<int>(TEST_KEY));
+        KeyValueStore.Instance.SetValue(TEST_KEY, 1337);
+        Assert.AreEqual(1337, KeyValueStore.Instance.GetValue<int>(TEST_KEY));
 
-
-            // Cleanup
-            KeyValueStore.DeleteKey(TEST_KEY);
-            Assert.IsNull(KeyValueStore.GetValue<int>(TEST_KEY));
-        }
+        // Cleanup
+        KeyValueStore.Instance.DeleteKey(TEST_KEY);
+        Assert.IsNull(KeyValueStore.Instance.GetValue<int?>(TEST_KEY));
     }
 }
